Add Keymap To Toggle Case Of Second Letter
For those annoying situations where you hold shift a little too long and
write HEllo instead of Hello. Now, with your cursor on the offending
word, you can go to Normal mode, then hit `<leader>tc` to fix the error and
bring the cursor to the end of the word and leave you in Normal mode.